Formatting fixes

+++ src/RandomNumberGenerator.sol
@@ -29,11 +29,10 @@ contract RandomNumberGenerator is VRFConsumerBaseV2, Ownable {
     event RandomNumberFulfilled(uint256 requestId, uint256[] randomWords);
     event BaccaratGameUpdated(address indexed oldGame, address indexed newGame);
 
-    constructor(
-        address _vrfCoordinator,
-        bytes32 _keyHash,
-        uint64 _subscriptionId
-    ) VRFConsumerBaseV2(_vrfCoordinator) Ownable(msg.sender) {
+    constructor(address _vrfCoordinator, bytes32 _keyHash, uint64 _subscriptionId)
+        VRFConsumerBaseV2(_vrfCoordinator)
+        Ownable(msg.sender)
+    {
         vrfCoordinator = VRFCoordinatorV2Interface(_vrfCoordinator);
         keyHash = _keyHash;
         subscriptionId = _subscriptionId;
@@ -57,17 +56,9 @@ contract RandomNumberGenerator is VRFConsumerBaseV2, Ownable {
     /// @notice Requests random numbers for the initial card deal
     /// @dev Requests 4 random words for the initial 4 cards
     /// @return requestId The ID of the VRF request
-    function requestRandomNumbers()
-        external
-        onlyBaccaratGame
-        returns (uint256)
-    {
+    function requestRandomNumbers() external onlyBaccaratGame returns (uint256) {
         uint256 requestId = vrfCoordinator.requestRandomWords(
-            keyHash,
-            subscriptionId,
-            REQUEST_CONFIRMATIONS,
-            CALLBACK_GAS_LIMIT,
-            INITIAL_NUM_WORDS
+            keyHash, subscriptionId, REQUEST_CONFIRMATIONS, CALLBACK_GAS_LIMIT, INITIAL_NUM_WORDS
         );
 
         requestIdToFulfilled[requestId] = false;
@@ -78,17 +69,9 @@ contract RandomNumberGenerator is VRFConsumerBaseV2, Ownable {
     /// @notice Requests random numbers for additional cards
     /// @dev Requests 2 random words for potential third cards
     /// @return requestId The ID of the VRF request
-    function requestAdditionalCards()
-        external
-        onlyBaccaratGame
-        returns (uint256)
-    {
+    function requestAdditionalCards() external onlyBaccaratGame returns (uint256) {
         uint256 requestId = vrfCoordinator.requestRandomWords(
-            keyHash,
-            subscriptionId,
-            REQUEST_CONFIRMATIONS,
-            CALLBACK_GAS_LIMIT,
-            ADDITIONAL_NUM_WORDS
+            keyHash, subscriptionId, REQUEST_CONFIRMATIONS, CALLBACK_GAS_LIMIT, ADDITIONAL_NUM_WORDS
         );
 
         requestIdToFulfilled[requestId] = false;
@@ -100,10 +83,7 @@ contract RandomNumberGenerator is VRFConsumerBaseV2, Ownable {
     /// @dev Processes random numbers and forwards them to the Baccarat game contract
     /// @param requestId The ID of the VRF request being fulfilled
     /// @param randomWords Array of random numbers received from VRF
-    function fulfillRandomWords(
-        uint256 requestId,
-        uint256[] memory randomWords
-    ) internal override {
+    function fulfillRandomWords(uint256 requestId, uint256[] memory randomWords) internal override {
         requestIdToFulfilled[requestId] = true;
         requestIdToRandomWords[requestId] = randomWords;
         emit RandomNumberFulfilled(requestId, randomWords);
@@ -111,14 +91,14 @@ contract RandomNumberGenerator is VRFConsumerBaseV2, Ownable {
         if (randomWords.length == INITIAL_NUM_WORDS) {
             // Initial deal
             uint8[4] memory cards;
-            for (uint i = 0; i < 4; i++) {
+            for (uint256 i = 0; i < 4; i++) {
                 cards[i] = uint8((randomWords[i] % 52) + 1);
             }
             IBaccaratGame(baccaratGame).receiveInitialCards(requestId, cards);
         } else {
             // Additional cards
             uint8[2] memory cards;
-            for (uint i = 0; i < 2; i++) {
+            for (uint256 i = 0; i < 2; i++) {
                 cards[i] = uint8((randomWords[i] % 52) + 1);
             }
             IBaccaratGame(baccaratGame).receiveThirdCard(requestId, cards);
@@ -128,9 +108,7 @@ contract RandomNumberGenerator is VRFConsumerBaseV2, Ownable {
     /// @notice Checks if a specific random number request has been fulfilled
     /// @param requestId The ID of the request to check
     /// @return bool True if the request has been fulfilled, false otherwise
-    function isRequestFulfilled(
-        uint256 requestId
-    ) external view returns (bool) {
+    function isRequestFulfilled(uint256 requestId) external view returns (bool) {
         return requestIdToFulfilled[requestId];
     }
 
@@ -138,9 +116,7 @@ contract RandomNumberGenerator is VRFConsumerBaseV2, Ownable {
     /// @param requestId The ID of the fulfilled request
     /// @return uint256[] Array of random numbers
     /// @dev Reverts if the request has not been fulfilled
-    function getRandomNumbers(
-        uint256 requestId
-    ) external view returns (uint256[] memory) {
+    function getRandomNumbers(uint256 requestId) external view returns (uint256[] memory) {
         require(requestIdToFulfilled[requestId], "Request not fulfilled");
         return requestIdToRandomWords[requestId];
     }
